\name{Game.of.Codes-package}
\alias{Game.of.Codes-package}
\alias{Game.of.Codes}
\docType{package}
\title{
\packageTitle{Game.of.Codes}
}
\description{
\packageDescription{Game.of.Codes}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Game.of.Codes}
\packageIndices{Game.of.Codes}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Game.of.Codes}

Maintainer: \packageMaintainer{Game.of.Codes}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
