% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game of Codes.R
\name{lexpectancy.allegiance}
\alias{lexpectancy.allegiance}
\title{Life expectancy by allegiance}
\usage{
lexpectancy.allegiance(x)
}
\arguments{
\item{x}{String: The name of the house}
}
\description{
Life expectancy by allegiance
}
\seealso{
\code{\link{mean}} which this function wraps
}

