% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game of Codes.R
\name{lexpectancy.gender}
\alias{lexpectancy.gender}
\title{Life expectancy by gender}
\usage{
lexpectancy.gender(x)
}
\arguments{
\item{x}{Dummy: 0 for female and 1 for male}
}
\description{
Life expectancy by gender
}
\seealso{
\code{\link{mean}} which this function wraps
}

