% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game of Codes.R
\name{lexpectancy.nobility}
\alias{lexpectancy.nobility}
\title{Life expectancy by nobility}
\usage{
lexpectancy.nobility(x)
}
\arguments{
\item{x}{Dummy: 0 for not noble and 1 for noble}
}
\description{
Life expectancy by nobility
}
\seealso{
\code{\link{mean}} which this function wraps
}

